+++ config/theme-supports.php
@@ -9,7 +9,7 @@
  * @since 	1.0.7
  * @license GPL-2.0+
  * @link    http://studiomoare.com/
- * @version 2.0.1
+ * @version 2.0.2
  */
 
 return [
@@ -28,7 +28,6 @@ return [
 	],
 	'genesis-menus'                   => [
 		'primary'     => __( 'Main Menu', 'moaregenesis' ),
-		'subfooter'   => __( 'Sub Footer Menu', 'moaregenesis' ),
 	],
 	'html5'                           => [
 		'caption',